backend/depannini/requirements.txt
Django==5.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
django-allauth==65.7.0
pillow==11.2.1
google-auth==2.39.0
channels==4.2.2
channels-redis==4.2.1
daphne==4.1.2
requests==2.32.3
twilio==9.6.1
python-dotenv==1.1.0
psycopg2-binary==2.9.10
whitenoise==6.9.0
dj-database-url==3.0.0